requirements.txt
fastapi == 0.46.0
investpy == 1.0.8
uvicorn == 0.11.1
pytest == 5.3.2
requests == 2.22.0
python-dotenv== 0.20.0
